requirements-test.txt
selenium>=4.16.0
webdriver-manager>=4.0.1
pytest-selenium>=4.0.0 